Tape class and TapeHeadDirection moved to Machine namespace.
