+++ src/TuringMachine/UnitTest/UnitTests/TransitionCollectionValidatorTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using TuringMachine.Machine;
 using TuringMachine.Transition;
 using Xunit;
